+++ src/test/scala/ai/privado/languageEngine/ruby/passes/MethodFullNameForInternalNodesTest.scala
@@ -11,13 +11,12 @@ import io.joern.x2cpg.X2Cpg.applyDefaultOverlays
 import io.joern.x2cpg.passes.frontend.{LocalKey, SBKey, SymbolTable}
 import io.shiftleft.semanticcpg.language._
 import ai.privado.utility.Utilities.resolver
-
-case class SourceCodeModel(sourceCode: String, fileName: String)
+import ai.privado.languageEngine.ruby.RubyTestBase._
 
 class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with BeforeAndAfterAll {
 
   "method full name for internal nodes belonging to the same module" should {
-    val cpg = code(
+    val (cpg, _) = code(
       List(
         SourceCodeModel(
           """
@@ -50,7 +49,8 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
         |""".stripMargin,
           "payment.rb"
         )
-      )
+      ),
+      applyPostProcessingPass = true
     )
 
     "have a resolved type when LHS is a call node and called without scopeResolution" in {
@@ -74,7 +74,7 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
 
   "method fullname propagation due to object creation with `new`" should {
 
-    val cpg = code(
+    val (cpg, _) = code(
       List(
         SourceCodeModel(
           """
@@ -102,7 +102,8 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
           |""".stripMargin,
           "demo.rb"
         )
-      )
+      ),
+      applyPostProcessingPass = true
     )
 
     "have correct type for the new node" in {
@@ -118,7 +119,7 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
 
   "types for nodes accessed via module" should {
 
-    val cpg = code(
+    val (cpg, _) = code(
       List(
         SourceCodeModel(
           """
@@ -141,7 +142,8 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
         |""".stripMargin,
           "demo.rb"
         )
-      )
+      ),
+      applyPostProcessingPass = true
     )
 
     "be generated" in {
@@ -150,7 +152,7 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
   }
 
   "Call named `perform` and `perform_async`" should {
-    val cpg = code(
+    val (cpg, _) = code(
       List(
         SourceCodeModel(
           """
@@ -177,7 +179,8 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
           |""".stripMargin,
           "demo.rb"
         )
-      )
+      ),
+      applyPostProcessingPass = true
     )
 
     "be resolved and linked to the same method `perform`" in {
@@ -191,29 +194,4 @@ class MethodFullNameForInternalNodesTest extends AnyWordSpec with Matchers with 
       performCall.name.l shouldBe List("perform", "perform_async")
     }
   }
-}
-
-def code(sourceCodes: List[SourceCodeModel]): Cpg = {
-  val inputDir = File.newTemporaryDirectory()
-  for (sourceCode <- sourceCodes) {
-    (inputDir / sourceCode.fileName).write(sourceCode.sourceCode)
-  }
-  val outputFile = File.newTemporaryFile()
-  val config = Config()
-    .withInputPath(inputDir.pathAsString)
-    .withOutputPath(outputFile.pathAsString)
-    .withUseDeprecatedFrontend(true)
-  val rubySrc = new RubySrc2Cpg()
-  val xtocpg = rubySrc.createCpg(config).map { cpg =>
-    applyDefaultOverlays(cpg)
-    cpg
-  }
-  val cpg               = xtocpg.get
-  val globalSymbolTable = new SymbolTable[LocalKey](SBKey.fromNodeToLocalKey)
-  new GlobalImportPass(cpg, globalSymbolTable).createAndApply()
-
-  new PrivadoRubyTypeRecoveryPassGenerator(cpg, globalSymbolTable).generate().foreach(_.createAndApply())
-  new RubyTypeHintCallLinker(cpg).createAndApply()
-
-  cpg
 }
