+++ src/test/scala/ai/privado/languageEngine/ruby/passes/SchemaParserTest.scala
@@ -1,25 +1,24 @@
-package ai.privado.languageEngine.ruby
+package ai.privado.languageEngine.ruby.passes
 
 import ai.privado.cache.RuleCache
+import ai.privado.languageEngine.ruby.RubyTestBase.*
 import ai.privado.languageEngine.ruby.passes.SchemaParser
-import io.shiftleft.codepropertygraph.generated.Cpg
-import org.scalatest.BeforeAndAfterAll
-import org.scalatest.matchers.should.Matchers
-import org.scalatest.wordspec.AnyWordSpec
+import ai.privado.semantic.Language.*
 import better.files.File
 import io.joern.rubysrc2cpg.deprecated.passes.RubyTypeHintCallLinker
 import io.joern.rubysrc2cpg.{Config, RubySrc2Cpg}
 import io.joern.x2cpg.X2Cpg.applyDefaultOverlays
 import io.joern.x2cpg.passes.frontend.{LocalKey, SBKey, SymbolTable}
+import io.shiftleft.codepropertygraph.generated.Cpg
 import io.shiftleft.semanticcpg.language.*
-import ai.privado.semantic.Language._
-
-case class SourceCodeModel(sourceCode: String, fileName: String)
+import org.scalatest.BeforeAndAfterAll
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
 
 class SchemaParserTest extends AnyWordSpec with Matchers with BeforeAndAfterAll {
 
   "Schema parser" should {
-    val cpg = code(
+    val (cpg, config) = code(
       List(
         SourceCodeModel(
           """
@@ -36,6 +35,7 @@ class SchemaParserTest extends AnyWordSpec with Matchers with BeforeAndAfterAll 
         )
       )
     )
+    new SchemaParser(cpg, config.inputPath, RuleCache()).createAndApply()
 
     "be able to create sql nodes" in {
       val table = cpg.sqlTable.l
@@ -50,24 +50,4 @@ class SchemaParserTest extends AnyWordSpec with Matchers with BeforeAndAfterAll 
     }
   }
 
-}
-
-def code(sourceCodes: List[SourceCodeModel]): Cpg = {
-  val inputDir = File.newTemporaryDirectory()
-  for (sourceCode <- sourceCodes) {
-    (inputDir / sourceCode.fileName).write(sourceCode.sourceCode)
-  }
-  val outputFile = File.newTemporaryFile()
-  val config = Config()
-    .withInputPath(inputDir.pathAsString)
-    .withOutputPath(outputFile.pathAsString)
-    .withUseDeprecatedFrontend(true)
-  val rubySrc = new RubySrc2Cpg()
-  val xtocpg = rubySrc.createCpg(config).map { cpg =>
-    applyDefaultOverlays(cpg)
-    cpg
-  }
-  val cpg = xtocpg.get
-  new SchemaParser(cpg, inputDir.pathAsString, new RuleCache).createAndApply()
-  cpg
 }
